+++ api/siteapp/views/role.py
@@ -1,3 +1,6 @@
+from django.db.models import Q
+from rest_framework import filters
+
 from api.base.views.base import SerializerClasses
 from api.base.views.viewsets import ReadWriteViewSet
 from api.siteapp.serializers.role import SimpleRoleSerializer, WriteRoleSerializer
@@ -6,10 +9,15 @@ from siteapp.models import Role
 
 class RoleViewSet(ReadWriteViewSet):
     queryset = Role.objects.all()
-
+    
+    search_fields = ['title']
+    filter_backends = (filters.SearchFilter,)
+    
     serializer_classes = SerializerClasses(retrieve=SimpleRoleSerializer,
                                            list=SimpleRoleSerializer,
                                            create=WriteRoleSerializer,
                                            update=WriteRoleSerializer,
                                            destroy=WriteRoleSerializer)
-    # filter_class = RoleFilter
+
+    def search(self, request, keyword):
+        return Q(title__icontains=keyword)