+++ frontend/src/index.js
@@ -6,5 +6,15 @@ import './components/projects/project-view/menu';
 import './components/permissions/component';
 import './components/permissions/permissions';
 
+import './components/requests/component';
+import './components/requests/requests';
+
+import './components/system-owner-approval/component';
+import './components/system-owner-approval/requireApprovalModal';
+import './components/system-owner-approval/proposal-steps';
+
+import './components/system-summary/component';
+import './components/system-summary/system_summary';
+
 import './components/cmpt_parties/component';
 import './components/cmpt_parties/cmpt_parties';